==> Data_Preparation/Step-1_Preparing_rentals_template.ipynb <==
# Preparing New Dataset
This notebook is meant to generate a new dataset (only for January 2024) based on *Citibike Trip data* that can be used to determine **bike rentals** based on a <u>fixed time resolution</u>. 

## Pre-processing the original data
Before preparing a new dataset, the original data is pre-processed to be suitable for creating a new dataset.

# Reading the data
import pandas as pd
complete_data = pd.read_csv("C:/Users/singh/Desktop/TUD (All Semesters)/Courses - Semester 5 (TU Dresden)/Research Task - Spatial Modelling/Datasets/Citibike Trip Data/202401-citibike-tripdata.csv")
complete_data.info()

# Types of ride vehicles
complete_data["rideable_type"].unique()

The dataset contains both *ebikes* and *classic bikes*. The complete dataset is filtered for **classic bikes** only since they are much more in numbers and is the primary source of mobility service.

# filtering the dataset
cbikes = complete_data[complete_data["rideable_type"] == "classic_bike"]
cbikes["rideable_type"].unique()

Additionally, there might be data points with <u>no information</u> about the **starting station name and location**. These data points can't provide information about rentals, hence they should be removed from the dataset.

# there are no rows with missing info about start
sum(cbikes["start_station_name"].isna())

# checking for redundant columns
cbikes.head()

The columns *rideable_type* and *member_casual* are redundant and can be removed from the dataset.

# reducing columns
cbikes = cbikes[["ride_id", "started_at", "ended_at", "start_station_name", "start_station_id", "end_station_name", "end_station_id", "start_lat", "start_lng", "end_lat", "end_lng"]]
cbikes.columns

# 'ride_id' can be used to count bike rentals
print(len(cbikes), len(cbikes["ride_id"].unique()))

## Creating New Dataset
The new dataset should contain stations with their location and time information included. Each row represents <u> one unit of chosen time resolution</u>.
<br><br>
**NOTE:** Station ID needs to be retained because there are some stations with the same name but slightly different locations!
<br><br>
In the dataset, some IDs are stored as text and others are stored as numbers. Converting th data type to *string* does not work because **\n** and some other text is being added automatically.
<br><br>
**Solution:**<br>
The stations can be manually given unique ids for the purpose of indentification.

# Creating the new dataset
rentals = pd.DataFrame()
rentals[["name", "lat", "lng"]] = cbikes[["start_station_name", "start_lat", "start_lng"]].drop_duplicates()
rentals = rentals.sort_values(by = ["name"], ignore_index=True)
rentals.head()

# Adding IDs manually
rentals["station_id"] = range(len(rentals))
rentals.head()

# Verifying the number of stations
print(len(cbikes[["start_station_name", "start_lat", "start_lng"]].drop_duplicates()), len(rentals["station_id"]))

# 5 stations have repeating names, but their location is different
rentals[rentals["name"].duplicated()]

# example
rentals[rentals["name"] == "Amsterdam Ave & W 119 St"]

## Altering the original (filtered) dataset
Instead of relying on station_ids provided by **Citibike**, the manually constructed IDs can be entered in dataset. This will make it easy for referencing and connecting *rentals* table with this original filteered dataset.<br><br>
<u>Note</u>: Original filtered dataset means data for classic bikes where NaN values and redundant columns are removed i.e. **cbikes** dataframe.

# Copying cbikes df
cbikes_altered = cbikes
cbikes_altered.head()

# Merging the two dfs
new_df = pd.merge(
    left=cbikes_altered, 
    right=rentals,
    how='left',
    left_on=['start_station_name', 'start_lat', 'start_lng'],
    right_on=['name', 'lat', 'lng'],
)

# Preserving bike trip info only with the starting station info
cbikes_altered = new_df[["ride_id", "started_at", "start_station_name", "station_id", "start_lat", "start_lng"]]
cbikes_altered.head()

# Checking validity
len(cbikes_altered["station_id"].unique())

Hence, the created dataset is correct.

## Exporting this dataset
This dataset can be used to calculate number of rentals by the manually generated station IDs.

# exporting to 'C:\Users\singh\Desktop\TUD (All Semesters)\Courses - Semester 5 (TU Dresden)\Research Task - Spatial Modelling\Code'
cbikes_altered.to_csv("Tripdata_with_manual_ids.csv", index=False)

## Adding time dimensionality to the *rentals* dataset
Now that station location information is available, a time resolution needs to be added so that number of rentals can be calculated in terms of that time unit, per station.

# Checking the range of temporal information in trip data
print(min(cbikes_altered["started_at"])) 
print(max(cbikes_altered["started_at"]))

# how is time info stored?
print(cbikes_altered["started_at"].iloc[0],
type(cbikes_altered["started_at"].iloc[0]))

# datetime info stored as a string
h = '2024-01-07 08:42:55.647'
print(h)
min(cbikes_altered["started_at"]) # without the print cmd

Date-time info will have to be extracted from the **str** class to *datetime* objects. 

## Generating 4 hr time resolution
Each day will be divided into four hour units. So there should be 6 units of time per day spanning as follows:<br>
- 0 to 4
- 4 to 8
- 8 to 12
- 12 to 16
- 16 to 20
- last 4 hours (which are assumed to be from 20:00 onwards)

# defining time units for a given date
time_res = ['00:00:00.000', '04:00:00.000', '08:00:00.000', '12:00:00.000', '16:00:00.000', '20:00:00.000']

# repeating time units 32 times
time_res = time_res*32
print(time_res[1], time_res[8])

# time operations can even be done in strings
'03:00:00.000' < '04:00:00.000'

# creating dates
dates_included = []

while len(dates_included) < 32:
    dates_included.append('2024-01-01')

print(dates_included[:3], len(dates_included))

nums = ['00', '01', '02','03','04','05','06','07','08','09','10','11','12','13','14','15','16','17','18','19','20','21','22','23','24','25','26','27','28','29','30','31']

# correcting days
for i in range(len(nums)):
    dates_included[i] = dates_included[i][:-2] + nums[i]

dates_included

dates_included[0] = '2023-12-31'
dates_included[:5]

Now, we are supposed to combine the date and the time information together. Each day has 6 time units, so each day must be repeated 6 times to accodomate the time resolution. However, there are actually 7 time windows because of 6 time units.

# repeating each element 6 times
dates_included = [element for element in dates_included for _ in range(6)]
dates_included

# Adding time dimension
date_time = dates_included

for t in range(0,len(dates_included)):
    date_time[t] = dates_included[t] + ' ' + time_res[t]

date_time[:6]

# for 1st January 2024
date_time[:14]

**Note:** The time <u>2023-12-31 00:00:00.000</u> means that bike rentals are mentioned between **00-04** hours. At every checkpoint the demand gets mentioned for the coming four hours.
<br><br>
The created date_time information must be added for each station in *rentals*.

## Finalizing the rentals dataset (without demand)

# rows needed for each station
len(date_time)

# duplicating each station 192 times
rentals = rentals.loc[rentals.index.repeat(192)].reset_index(drop=True)
rentals.head()

# Repeating date_time: 2124 unique stations
date_time = date_time*2124
rentals["datetime"] = date_time
rentals.head()

len(rentals)

# exporting the rentals dataset
rentals.to_csv("rentals_without_demand.csv", index=False)
