==> Data_Preparation/Step-2_Calculating_rental_demand.ipynb <==
# Updating rental demand 
This notebook uses the datasets previously created to determine bike rentals for each station, for every time unit (i.e. six *4 hour* units). The dataset is being created only for the month of January 2024.

## Loading the datasets

# Importing the datasets
import pandas as pd
rentals = pd.read_csv("C:/Users/singh/Desktop/TUD (All Semesters)/Courses - Semester 5 (TU Dresden)/Research Task - Spatial Modelling/Code/rentals_without_demand.csv")
tripdata = pd.read_csv("C:/Users/singh/Desktop/TUD (All Semesters)/Courses - Semester 5 (TU Dresden)/Research Task - Spatial Modelling/Code/Tripdata_with_manual_ids.csv")

# Overview of the data
rentals.head()

# Overview of the data
tripdata.head()

# adding a rentals column 
rentals['#_rentals'] = 0
rentals.head()

# demand calculation for 31 Dec'23 (00-04 hrs)
tripdata[(tripdata["started_at"] < '2023-12-31 04:00:00.000') & (tripdata["started_at"] > '2023-12-31 00:00:00.000')]

There is only 1 rental for the 1st time unit for 31 Dec 2023. The rental is only made for <u>Broadway & W 36 St</u> station.

## Adding datetime attributes
To update demand (i.e. rental information), I will need to add date and time information to both the *trip dataset* as well as the *rentals* dataset.

pd.to_datetime('2023-12-31 23:59:59.999').time().microsecond

# Updating the rentals dataset
rentals['year'] = pd.Series([x.date().year for x in pd.to_datetime(rentals["datetime"])])
rentals['month'] = pd.Series([x.date().month for x in pd.to_datetime(rentals["datetime"])])
rentals['day'] = pd.Series([x.date().day for x in pd.to_datetime(rentals["datetime"])])

rentals['hour'] = pd.Series([x.time().hour for x in pd.to_datetime(rentals["datetime"])])
rentals['minute'] = pd.Series([x.time().minute for x in pd.to_datetime(rentals["datetime"])])
rentals['second'] = pd.Series([x.time().second for x in pd.to_datetime(rentals["datetime"])])
rentals['microsecond'] = pd.Series([x.time().microsecond for x in pd.to_datetime(rentals["datetime"])])

# Updating the trips dataset
tripdata['started_at_year'] = pd.Series([x.date().year for x in pd.to_datetime(tripdata["started_at"])])
tripdata['started_at_month'] = pd.Series([x.date().month for x in pd.to_datetime(tripdata["started_at"])])
tripdata['started_at_day'] = pd.Series([x.date().day for x in pd.to_datetime(tripdata["started_at"])])

tripdata['started_at_hour'] = pd.Series([x.time().hour for x in pd.to_datetime(tripdata["started_at"])])
tripdata['started_at_minute'] = pd.Series([x.time().minute for x in pd.to_datetime(tripdata["started_at"])])
tripdata['started_at_second'] = pd.Series([x.time().second for x in pd.to_datetime(tripdata["started_at"])])
tripdata['started_at_microsecond'] = pd.Series([x.time().microsecond for x in pd.to_datetime(tripdata["started_at"])])

# rentals updated
rentals.head()

# tripdata updated
tripdata.head()

## Calculating rentals
With date and time attributes, determining rentals should be doable.

# This is how to modify individual elements!
rentals.loc[(rentals["station_id"] == 656) & (rentals["year"] == 2023) & (rentals["hour"] == 0),"#_rentals"] = 0

# determining demand for 2023; 00-04hrs
for i in rentals["station_id"].unique():
    idx = (rentals["station_id"] == i) & (rentals["year"] == 2023) & (rentals["hour"] == 0)
    rentals.loc[idx, "#_rentals"] = len(tripdata[(tripdata["started_at_year"] == 2023) & (tripdata["started_at_hour"] < 4) & (tripdata["station_id"] == i)])

# determining demand for 2023; 04-08hrs
for i in rentals["station_id"].unique():
    idx = (rentals["station_id"] == i) & (rentals["year"] == 2023) & (rentals["hour"] == 4)
    rentals.loc[idx, "#_rentals"] = len(tripdata[(tripdata["started_at_year"] == 2023) & (tripdata["started_at_hour"] < 8) & (tripdata["started_at_hour"] >= 4) & (tripdata["station_id"] == i)])

# determining demand for 2023; 08-12hrs
for i in rentals["station_id"].unique():
    idx = (rentals["station_id"] == i) & (rentals["year"] == 2023) & (rentals["hour"] == 8)
    rentals.loc[idx, "#_rentals"] = len(tripdata[(tripdata["started_at_year"] == 2023) & (tripdata["started_at_hour"] < 12) & (tripdata["started_at_hour"] >= 8) & (tripdata["station_id"] == i)])

# determining demand for 2023; 12-16hrs
for i in rentals["station_id"].unique():
    idx = (rentals["station_id"] == i) & (rentals["year"] == 2023) & (rentals["hour"] == 12)
    rentals.loc[idx, "#_rentals"] = len(tripdata[(tripdata["started_at_year"] == 2023) & (tripdata["started_at_hour"] < 16) & (tripdata["started_at_hour"] >= 12) & (tripdata["station_id"] == i)])

# determining demand for 2023; 16-20hrs
for i in rentals["station_id"].unique():
    idx = (rentals["station_id"] == i) & (rentals["year"] == 2023) & (rentals["hour"] == 16)
    rentals.loc[idx, "#_rentals"] = len(tripdata[(tripdata["started_at_year"] == 2023) & (tripdata["started_at_hour"] < 20) & (tripdata["started_at_hour"] >= 16) & (tripdata["station_id"] == i)])

# determining demand for 2023; 20-00hrs
for i in rentals["station_id"].unique():
    idx = (rentals["station_id"] == i) & (rentals["year"] == 2023) & (rentals["hour"] == 20)
    rentals.loc[idx, "#_rentals"] = len(tripdata[(tripdata["started_at_year"] == 2023) & (tripdata["started_at_hour"] >= 20) & (tripdata["station_id"] == i)])

# verifying rental information for 31 Dec 2023
sum(rentals.loc[(rentals["year"] == 2023) & (rentals["#_rentals"] > 0),"#_rentals"])

# the rental demand is correct!
len(tripdata[(tripdata["started_at_year"] == 2023)])

> The rentals have been successfully determined for 2023. There was only one day included in this year i.e. *31 Dec* so calculating rentals was not difficult here.

# For 2024, data is only present for January
max(rentals.loc[rentals["year"] == 2024,"month"])

# Updating for 2024
for k in range(1, 32):
    
    # determining demand for 2024; 00-04hrs
    for i in rentals["station_id"].unique():
        idx = (rentals["station_id"] == i) & (rentals["year"] == 2024) & (rentals["hour"] == 0) & (rentals["day"] == k)
        rentals.loc[idx, "#_rentals"] = len(tripdata[(tripdata["started_at_year"] == 2024) & (tripdata["started_at_day"] == k) & (tripdata["started_at_hour"] < 4) & (tripdata["station_id"] == i)])
        
    # determining demand for 2024; 04-08hrs
    for i in rentals["station_id"].unique():
        idx = (rentals["station_id"] == i) & (rentals["year"] == 2024) & (rentals["hour"] == 4) & (rentals["day"] == k)
        rentals.loc[idx, "#_rentals"] = len(tripdata[(tripdata["started_at_year"] == 2024) & (tripdata["started_at_day"] == k) & (tripdata["started_at_hour"] >= 4) & (tripdata["started_at_hour"] < 8) & (tripdata["station_id"] == i)])
        
    # determining demand for 2024; 08-12hrs
    for i in rentals["station_id"].unique():
        idx = (rentals["station_id"] == i) & (rentals["year"] == 2024) & (rentals["hour"] == 8) & (rentals["day"] == k)
        rentals.loc[idx, "#_rentals"] = len(tripdata[(tripdata["started_at_year"] == 2024) & (tripdata["started_at_day"] == k) & (tripdata["started_at_hour"] >= 8) & (tripdata["started_at_hour"] < 12) & (tripdata["station_id"] == i)])
        
    # determining demand for 2024; 12-16hrs
    for i in rentals["station_id"].unique():
        idx = (rentals["station_id"] == i) & (rentals["year"] == 2024) & (rentals["hour"] == 12) & (rentals["day"] == k)
        rentals.loc[idx, "#_rentals"] = len(tripdata[(tripdata["started_at_year"] == 2024) & (tripdata["started_at_day"] == k) & (tripdata["started_at_hour"] >= 12) & (tripdata["started_at_hour"] < 16) & (tripdata["station_id"] == i)])
        
    # determining demand for 2024; 16-20hrs
    for i in rentals["station_id"].unique():
        idx = (rentals["station_id"] == i) & (rentals["year"] == 2024) & (rentals["hour"] == 16) & (rentals["day"] == k)
        rentals.loc[idx, "#_rentals"] = len(tripdata[(tripdata["started_at_year"] == 2024) & (tripdata["started_at_day"] == k) & (tripdata["started_at_hour"] >= 16) & (tripdata["started_at_hour"] < 20) & (tripdata["station_id"] == i)])
        
    # determining demand for 2024; 20-24hrs
    for i in rentals["station_id"].unique():
        idx = (rentals["station_id"] == i) & (rentals["year"] == 2024) & (rentals["hour"] == 20) & (rentals["day"] == k)
        rentals.loc[idx, "#_rentals"] = len(tripdata[(tripdata["started_at_year"] == 2024) & (tripdata["started_at_day"] == k) & (tripdata["started_at_hour"] >= 20) & (tripdata["station_id"] == i)])

# total trips on jan 1 2024
len(tripdata[(tripdata["started_at_year"] == 2024)&(tripdata["started_at_day"] == 1)])

# rentals for jan 1 2024 is correct!
sum(rentals.loc[(rentals["year"] == 2024) & (rentals["day"] == 1),"#_rentals"])

## Exporting the dataset
The rentals dataset with complete demand information is fully prepared and can be used for further analysis.

# exporting
rentals.to_csv("rentals_with_demand.csv", index=False)

## Demand patterns by time
This section is meant to explore which parts of the day have higher rental demand. This is being done to cut out some parts of time units where rental demand is consistently low throught the month.

# importing
import pandas as pd
rentals = pd.read_csv("C:/Users/singh/Desktop/TUD (All Semesters)/Courses - Semester 5 (TU Dresden)/Research Task - Spatial Modelling/Code/rentals_with_demand.csv")
rentals.head()

Since minute, second and microsecond are not required they're being removed.

# dropping columns
rentals.drop(columns = ['minute','second','microsecond'], inplace=True)
rentals.columns

# available time units
rentals["hour"].unique()

# Average number of rentals
import statistics as st
means = []
for i in rentals["hour"].unique():
    means.append(st.mean(rentals.loc[rentals["hour"] == i, "#_rentals"]))

means

The average is taken for all days and all stations.

# The data spread
rentals[rentals["hour"] == 8]

From the means, it can be seen that most of the traffic is coming from the 3 following time units:
- 8 to 12
- 12 to 16
- 16 to 20

Since the rental volume is also somewhat cosniderable from *20 to 24*, bike traffic from *20 to 22* should also be considered since this time window is likely to contain most of the rentals from the time unit: <u>20 to 24</u>.
<br><br>
Hence, the time-frame per day that should be taken into account to accomodate majority of the rentals ranges from **8 AM - 10 PM**.

## Proposed New Time Units
A new dataset calculating rentals can be prepared with the following 6 time units, each of *two hour* gaps:
- 8 - 10
- 10 - 12
- 12 - 14
- 14 - 16
- 16 - 20
- 20 - 22
