==> Data_Preparation/Step-7_Combining_3_month_data.ipynb <==
# Training and Test Data
This notebook checks and prepares the final training and test data for the research task.
## Quality Check
This section is meant to inspect the quality of the created dataset and search for any discrepencies.

import pandas as pd
rentals_jan = pd.read_csv(r"C:\Users\singh\Desktop\TUD (All Semesters)\Courses - Semester 5 (TU Dresden)\Research Task - Spatial Modelling\Code\rentals_with_demand_new_time_units.csv")
rentals_feb = pd.read_csv(r"C:\Users\singh\Desktop\TUD (All Semesters)\Courses - Semester 5 (TU Dresden)\Research Task - Spatial Modelling\Code\rentals_with_demand_new_time_units_feb.csv")
rentals_mar = pd.read_csv(r"C:\Users\singh\Desktop\TUD (All Semesters)\Courses - Semester 5 (TU Dresden)\Research Task - Spatial Modelling\Code\rentals_with_demand_new_time_units_mar.csv")
rentals_apr = pd.read_csv(r"C:\Users\singh\Desktop\TUD (All Semesters)\Courses - Semester 5 (TU Dresden)\Research Task - Spatial Modelling\Code\rentals_with_demand_new_time_units_apr.csv")

# Columns preserved
rentals_jan.columns

# Unique stations in each dataset
print(len(rentals_jan["station_id"].unique()), len(rentals_feb["station_id"].unique()), len(rentals_mar["station_id"].unique()), len(rentals_apr["station_id"].unique()))

- March seems to have the most stations
- March datset has <u>3 additional stations</u>. This means that *station_id* is not uniform and is different for the month of **March**. This is also an issue for **April**, since there is <u>one additional station</u>.

# evaluating the time range for each dataset
print(min(rentals_jan["month"]), min(rentals_feb["month"]), min(rentals_mar["month"]), min(rentals_apr["month"]))
print(max(rentals_jan["month"]), max(rentals_feb["month"]), max(rentals_mar["month"]), max(rentals_apr["month"]))

# Even in rentals_jan, data is only stored for Jan (in addition to 31 Dec'23)
rentals_jan[rentals_jan["month"] == 2]

Data for timeline range looks good!

## Reading station information
Each dataset from January to April may not have the same station name and IDs. This looks especially true for March and April.

# finding elements in A not in B
a = [1,2,3,4]
b = [1,4]

set(a) - set(b)

# Comparing January data with remaining months
print(len(set(rentals_jan["name"]) - set(rentals_feb["name"])))
print(len(set(rentals_jan["name"]) - set(rentals_mar["name"])))
print(len(set(rentals_jan["name"]) - set(rentals_apr["name"])))

The month of january seems to have a lot of stations that were later discontinued.

# Comparing February data with remaining months
print(len(set(rentals_feb["name"]) - set(rentals_jan["name"])))
print(len(set(rentals_feb["name"]) - set(rentals_mar["name"])))
print(len(set(rentals_feb["name"]) - set(rentals_apr["name"])))

# Comparing March data with remaining months
print(len(set(rentals_mar["name"]) - set(rentals_jan["name"])))
print(len(set(rentals_mar["name"]) - set(rentals_feb["name"])))
print(len(set(rentals_mar["name"]) - set(rentals_apr["name"])))

# Number of common stations by name - all datasets
jan_stations = set(rentals_jan["name"].unique())
feb_stations = set(rentals_feb["name"].unique())
mar_stations = set(rentals_mar["name"].unique())
apr_stations = set(rentals_apr["name"].unique())
common_stations = set.intersection(jan_stations, feb_stations, mar_stations, apr_stations)
len(common_stations)

# Viewing the list
common_stations = list(common_stations)
common_stations[:5]

So, there are 2088 common stations throughout the entire four-month dataset. This must be used for data uniformity. Since there are some stations missing in each month dataset, the stored *station ids* are <u>not uniform</u>. 

# Comparing common records (for Jan) with Jan records
print(len(rentals_jan))
print(len(rentals_jan[rentals_jan["name"].isin(common_stations)]))

## Combining the dataset 
The entire 4 month dataset is combined into a single dataframe, by row.

# combining the data
citibike = pd.concat([rentals_jan, rentals_feb, rentals_mar, rentals_apr], axis=0, ignore_index=True)
citibike.info()

# dropping the station_id column since it is irrelevant
citibike.drop(columns=["station_id"], inplace=True)
citibike.columns

# rows after filtering for common stations
print(len(citibike))
len(citibike[citibike["name"].isin(common_stations)])

# filtering for common stations 
citibike = citibike[citibike["name"].isin(common_stations)]
len(citibike)

# number of unique stations
print(len(citibike["name"].unique()))

## Preparing station id
In this dataset, a station may have the same name but slightly different coordinates. So station *name* must be **combined** with the *coordinates* to get a unique <u>station id</u>. 

# Unique name-coordinate pair
station_identifier = citibike.drop_duplicates(subset=["name", "lat", "lng"])
print(len(station_identifier))
print(len(citibike["name"].unique()))

Hence, there are some stations in *citibike* with the same name but different coordinates!

# Preparing ID list
ids = citibike.drop_duplicates(subset=["name", "lat", "lng"], ignore_index=True)[["name", "lat", "lng"]]
ids.sort_values(by = "name", ignore_index=True, inplace=True)
ids.head()

# Assigning IDs
ids["ID"] = range(0,len(ids))
ids.head()

# Merging ids with citibike data
citibike = citibike.merge(ids, on=["name", "lat", "lng"], how='left')
citibike.head()

# Processing columns
citibike.drop(columns=["ID_x"], inplace=True)
citibike = citibike.rename(columns={'ID_y': 'ID'})
citibike.sort_values(by="name", ignore_index=True, inplace=True)
citibike.head()

# Cheking data for station: '1 Ave & E 110 St'
len(citibike[citibike["name"] == '1 Ave & E 110 St'])

# total records for station "1 Ave & E 110 St"
(7*32) + (7*29) + (7*31) + (7*30)

There should be 854 records for every station present in the dataset.

# for which station do we not have 854 rows?
records = []
for i in range(0,len(ids)):
    records.append(len(citibike[citibike["ID"] == i]))

records[:5]

# there are 8 stations having less records!
sum([j < 854 for j in records])

It appears that dataset is incomplete for 8 stations!

# Index: station ids
[i for i, num in enumerate(records) if num < 854]

# Checking for station ID: 504
len(citibike[citibike["ID"] == 504])

## Removing some stations
Stations with IDs *504, 580, 846, 1028, 1310, 1759, 1938, 2019* are being removed from the dataset since complete information for these stations is not available. Retaining these stations will lead to unnecessary loss in predicition performance.

# filtering the data
unwanted_ids = [504, 580, 846, 1028, 1310, 1759, 1938, 2019]
citibike = citibike[~citibike["ID"].isin(unwanted_ids)]
citibike.sort_values(by="name", ignore_index=True, inplace=True)
citibike[citibike["ID"] == 504]

# Unique station remaining
print(len(citibike["ID"].unique()))
print(len(citibike))

# Calculating total records now
854*2088

Hence, the dataset is correct and complete!

# Exporting the final data
citibike.to_csv("RENTALS.csv", index=False)
