==> Data_Preparation/Step-8_Identifying_200_nearest_stations.ipynb <==
## Station Locations to POINTs
The 4 month combined dataset is very huge. Instead of taking the entire, we will firstly take a look at the 200 nearest stations for a selected station. Since, the first station in the dataset is <u>1 Ave & E 110 St</u> the nearest 200 stations will be calculated.

# Importing the rentals dataset
import geopandas as gpd
import pandas as pd
rentals = pd.read_csv(r"C:\Users\singh\Desktop\TUD (All Semesters)\Courses - Semester 5 (TU Dresden)\Research Task - Spatial Modelling\Code\RENTALS.csv")
rentals.info()                      

# adding POINT geometry to rentals
rentals["coordinates"] = gpd.points_from_xy(rentals.lng, rentals.lat, crs="EPSG:4326")
rentals.head()

# Identifying uniques stations
stations = rentals.drop_duplicates(subset=['name', 'ID', 'coordinates'], ignore_index=True)
stations = stations[["name", "ID", "coordinates"]]
stations.head()

## Selecting Stations
Stations need to be filtered so that only the nearest stations are listed. Using stations that are closer together will help while modelling <u>spatial dependencies</u>. For determining which stations are closer, distance needs to be calculated. 

# stations df is not a GeoDataFrame
print(type(stations))

# Convert to a GeoDataFrame
stations_gdf = gpd.GeoDataFrame(stations, geometry='coordinates', crs="EPSG:4326")
print(type(stations_gdf))

# Point geometries
stations_gdf.head()

# Distance between the first two stations (shortest distance) 
stations_gdf.loc[0,"coordinates"].distance(stations_gdf.loc[1,"coordinates"])

# Calculating distance of each station with '1 Ave & E 110 St'
dist = []
for i in stations_gdf["coordinates"]:
    dist.append(stations_gdf.loc[0,"coordinates"].distance(i))
dist[:5]

# Adding dist as a column
stations_gdf["distance_to_1Ave&E110_St"] = dist
stations_gdf.head()

# nearest 200 stations
selected_stations = stations_gdf.sort_values(by="distance_to_1Ave&E110_St")
selected_stations = selected_stations.head(200)
selected_stations.head()

# filtering rentals data for chosen stations
selected_rentals = rentals[rentals["ID"].isin(list(selected_stations["ID"]))]
selected_rentals.info()

# Checking the number of stations
len(selected_rentals["ID"].unique())

# exporting the selected rentals
selected_rentals.to_csv("rentals_near200_st.csv", index=False)
