==> Data_Preparation/Step-9_Adding_weather_features.ipynb <==
# Predictions on Daily Rentals
The aim of this notebook is to see if there is any improvement in prediction performance by adding more weather related features i.e. temperature, wind, rainfall, etc.

## Creating a temperature dataset
The data is taken from the [National Weather Service](https://www.weather.gov/wrh/climate?wfo=okx). As a reference location for New York City, weather readings from **Central Park** station have been used.

# importing weather data
import pandas as pd
temp_daily = pd.DataFrame()
temp_daily.head()

# adding values for January 2024
day = [i for i in range(1,32)]
temp_daily["day"] = day
temp_daily["year"] = 2024
temp_daily["month"] = 1
temp_daily["avg_temp_(F)"] = 0
temp_daily.head()

# creating avg_temp column
temp = [41.0,35.5,38.5,36.5,31.5,34.5,36.0,40.5,46.5,50.5,44.0,45.0,47.5,35.0,26.0,27.0,20.5,28.0,29.0,22.0,25.5,31.5,37.0,43.0,52.5,44.0,46.0,40.5,38.0,36.5,37.0]
temp_daily["avg_temp_(F)"] = temp
temp_daily.head()

# creating precipitation column
T = 0.001 

prec = [0.03, 0.00, 0.00, 0.00, 0.00, 0.41, 0.24, 0.00, 1.73, 0.22, 0.00, 0.08, 0.81, T,
0.04, 0.28, 0.00, T,
0.04, 0.00, 0.00, 0.00, 0.05, 0.05, 0.24, 0.19, 0.00, 0.82, 0.05, T, T]

T means "Trace". It suggests that the prec was slightly above 0.00, but it can't be measured. Here the value is chosen to be *0.001* to <u>avoid missing values</u>.

temp_daily["prec"] = prec
temp_daily.head()

# creating snow depth column
sd = [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0]
temp_daily["snow_depth"] = sd
temp_daily.head()

As of now, the data has been added only for January 2024. The data for 31 Dec 2023 also needs to be added for the purpose of completion.

# How the data is stored
temp_daily.iloc[30]

# Updating
new_row = pd.DataFrame({'day':31, 'year':2023, 'month':12, 'avg_temp_(F)':41.0, 'prec':0.00, 'snow_depth':0},index=['31'])
temp_daily = pd.concat([new_row,temp_daily.loc[:]]).reset_index(drop=True)
temp_daily.info()

# It works!
print(temp_daily.iloc[0])
print(len(temp_daily))

## Combining weather with rentals_daily
Weather based features are combined with daily rentals dataset.

# importing
rentals_daily = pd.read_csv("C:/Users/singh/Desktop/TUD (All Semesters)/Courses - Semester 5 (TU Dresden)/Research Task - Spatial Modelling/Code/rentals_with_daily_demand.csv")

rentals_daily[(rentals_daily["day"] == 1)]

# merging the dfs
rentals_daily_temp = rentals_daily.merge(temp_daily, how='left', on= ['day','month','year'])

# It works!
rentals_daily_temp[(rentals_daily_temp["day"] == 31)&(rentals_daily_temp["station_id"] == 656)]

## Developing training and test sets
The default of 450 random stations are taken as before.

# ['name', 'lat', 'lng'] variables are excluded 
## no 'time' column is used

# training data: response variable
X_train = rentals_daily_temp.loc[(rentals_daily_temp["year"] == 2023) | ((rentals_daily_temp["year"] == 2024) & (rentals_daily_temp["day"] < 26)),["station_id", "year", "month", "day", "avg_temp_(F)", "prec", "snow_depth"]]

# test data: response variable
X_test = rentals_daily_temp.loc[(rentals_daily_temp["year"] == 2024) & (rentals_daily_temp["day"] >= 26),["station_id", "year", "month", "day", "avg_temp_(F)", "prec", "snow_depth"]]

import random
random.seed(42)
station_ids = list(rentals_daily_temp["station_id"].unique())

# Picking 450 unique station ids randomly
sel_st = random.sample(station_ids, 450)
sel_st[:5]

# filtering for 450 chosen stations
X_train = X_train[X_train["station_id"].isin(sel_st)]
X_test = X_test[X_test["station_id"].isin(sel_st)]

# creating dummies for X_train and X_test
X_train = pd.get_dummies(X_train, columns=['station_id'])
X_test = pd.get_dummies(X_test, columns=['station_id'])

# training data: target
y_train = rentals_daily_temp.loc[(rentals_daily_temp["year"] == 2023) | ((rentals_daily_temp["year"] == 2024) & (rentals_daily_temp["day"] < 26)), "rentals"]

# test data: target
y_test = rentals_daily_temp.loc[(rentals_daily_temp["year"] == 2024) & (rentals_daily_temp["day"] >= 26), "rentals"]

# Also need to filter y_train and y_test
y_train = y_train[y_train.index.isin(X_train.index)]
y_test = y_test[y_test.index.isin(X_test.index)]

## Applying a linear model
It turns out that a simple Linear Regression Model is actually performing better than a *random forest* for prediction performance on test set. This suggests that the relationship between features and the response variable is <u>not far from linear</u>.

# Checking for number of stations considered
X_train.columns

# training performance
from sklearn.linear_model import LinearRegression
lr = LinearRegression().fit(X_train, y_train)

from sklearn.metrics import mean_squared_error
from sklearn.metrics import r2_score
y_train_pred_lr = lr.predict(X_train)
print(mean_squared_error(y_train, y_train_pred_lr),r2_score(y_train, y_train_pred_lr))

# performance on test set
y_test_pred_lr = lr.predict(X_test)
print(mean_squared_error(y_test, y_test_pred_lr), r2_score(y_test, y_test_pred_lr))

## Applying random forest
This time the model has numeric weather attributes. The expectation is that this should lead to a performance increase.

# training the model
from sklearn.ensemble import RandomForestRegressor
forest = RandomForestRegressor(random_state=2) 
forest.fit(X_train, y_train)

# training performance
from sklearn.metrics import mean_squared_error
from sklearn.metrics import r2_score
y_train_pred = forest.predict(X_train)
print(mean_squared_error(y_train, y_train_pred),r2_score(y_train, y_train_pred))

# performance on test set
y_test_pred = forest.predict(X_test)
print(mean_squared_error(y_test, y_test_pred), r2_score(y_test, y_test_pred))

### Preliminary Results
#### Random Forest
Comparing the performance for test data as the number of stations are increased, the prediction accuracy seems to increase for <u>daily rentals</u>. 
- Switching to daily demand, the r2 score was 0.77 (50 stations)
- Switching to daily demand, the r2 score was 0.7 (450 stations)
- Adding weather information, the r2 score was 0.7 (450 stations)

Adding **weather attributes** increases the performance further for *daily rentals*, but not by a huge margin. Maybe the effect of this will be more visible when data is considered for 3-4 months.
<br><br>
**Note**: The <u>training time</u> seems to have *improved*. This was not expected.

#### Linear Regression
<u>Case for 450 stations</u>: Linear model has lower r2 score on training but better r2 score on the test set, compared to *random forest*. This suggests that random forest is **overfitting**.

### Further Actions
- With 50 stations: The r2-score of *linear model* is 0.79 
- With 450 stations: The r2-score of *linear model* is 0.77

It looks like the performance of linear model is decreasing with increase in the volume of data. Also, the performance of *random forest* seems to be changing every time it is run. This suggest that there might be some room for <u>hyperparameter tuning</u>.
